configuration-management/examples/python/pyproject.toml
[project]
name = "config-loader"
version = "1.0.0"
description = "Configuration loader for multi-environment configuration management"
requires-python = ">=3.10"

dependencies = [
    "pyyaml>=6.0",
    "google-cloud-secret-manager>=2.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
